requirements.txt
Django
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg[binary]
python-decouple==3.8
Pillow>=10.4.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0